requirements.txt
awscli==1.16.162
boto3==1.9.134
flask==1.0.2
flask-sqlalchemy  == 2.3.2
numpy==1.16.3
pandas==0.24.2
PyMySQL==0.9.3
pytest==4.0.1
python-dateutil==2.8.0
pyyaml==3.13
requests==2.21.0
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.3.0
six==1.12.0
SQLAlchemy==1.3.1
xgboost==0.82

